# app-java/src/main/resources/META-INF/resources/swagger-ui/openapi-order-v1-generated.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Order V1 API",
    "version" : "1.0.0"
  },
  "paths" : {
    "/orders/{orderId}" : {
      "get" : {
        "tags" : [ "Orders" ],
        "operationId" : "getOrder",
        "parameters" : [ {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrderDTO"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Orders" ],
        "operationId" : "postOrder",
        "parameters" : [ {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewOrderDTO"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrderDTO"
                }
              }
            }
          }
        }
      }
    },
    "/orders" : {
      "get" : {
        "tags" : [ "Orders" ],
        "operationId" : "getOrders",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/OrderDTO"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "OrderDTO" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string",
            "enum" : [ "ORDER_PLACED", "ORDER_IN_PROGRESS", "ORDER_CANCELED", "ORDER_COMPLETED" ]
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/OrderItemDTO"
            }
          }
        }
      },
      "OrderItemDTO" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          }
        }
      },
      "NewOrderDTO" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      }
    }
  }
}